Login Terminado

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserServerSession } from '@/app/auth/actions/auth-actions'
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import { boolean, object, string } from 'yup'
@@ -31,11 +32,16 @@ const postSchema = object({
 
 export async function POST(request: Request) { 
 
+  const user = await getUserServerSession()
+  if(!user) {
+    return NextResponse.json('No autorizado', {status: 401})
+  }
+
   try {
 
     const {complete, description} = await postSchema.validate(await request.json()) 
 
-    const todo = await prisma.todo.create({data: {complete, description}})
+    const todo = await prisma.todo.create({data: {complete, description, userId: user.id as any}})
 
     return NextResponse.json(todo)
     
@@ -47,9 +53,14 @@ export async function POST(request: Request) {
 }
 export async function DELETE(request: Request) { 
 
+  const user = await getUserServerSession()
+  if(!user) {
+    return NextResponse.json('No autorizado', {status: 401})
+  }
+
   try {
 
-    await prisma.todo.deleteMany({where: {complete: true}})
+    await prisma.todo.deleteMany({where: {complete: true, userId: user.id}})
 
     return NextResponse.json('Borrados')
